CloudFront invalidation.
AWS doesn’t provide a command line tool for CloudFront. Therefore, if you want to invalidate objects, you’ll need to log into the console and add the objects to invalidate. It’s a really long and tiresome process. Instead, you can make use of the Boto library.
This script reads objects specified in a file, invalidates them, prints an invalidation ID, and clears the contents of invalid.txt.